Exposing message compression from DNS library
When facing large messages, one can have its dns refusing the update due
to "too large message" error.
The DNS library used propose to compress the message that is sent.

This commit exposes this feature. It adds a new command line option
named --enable-compression and adds a new field to the configuration.
The default value of the field is set to false to not change the current
behavior of external-dns.